app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.1
google-api-python-client==2.108.0
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0